requirements.txt
attrs==22.1.0
click==8.1.3
coverage==6.5.0
exceptiongroup==1.0.4
Flask==2.2.2
importlib-metadata==5.1.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==22.0
Pillow==9.3.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pytest==7.2.0
python-dotenv==0.21.0
tomli==2.0.1
Werkzeug==2.2.2
zipp==3.11.0
Flask-Mail==0.9.1
